refactor(types): extract diff receipt DTOs
