+++ ci/proof.toml
@@ -1122,6 +1122,7 @@ paths = [
   "crates/tokmd/schemas/**",
   "crates/tokmd/tests/schema*.rs",
   "crates/tokmd-types/src/context.rs",
+  "crates/tokmd-types/src/diff.rs",
   "crates/tokmd-types/src/lib.rs",
   "crates/tokmd-types/tests/schema*.rs",
   "docs/SCHEMA.md",
